requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
anthropic
python-docx
pdfplumber
Pillow
pandas
openpyxl
reportlab
python-dotenv
python-multipart
